1) Add dynamically added/removed input fields functionality 2)Add validation 3)Edit BookController and BookViewModel 4)Add AuthorViewModel 5)Add default image for book in wwwroot

+++ LvivCompany.Bookstore/LvivCompany.Bookstore.Entities/Models/Book.cs
@@ -14,7 +14,7 @@ namespace LvivCompany.Bookstore.Entities
 
         public string Description { get; set; }
 
-        public short Year { get; set; }
+        public int Year { get; set; }
 
         public long PublisherId { get; set; }
